+++ cli/cli/commands/sandbox/sandbox.go
@@ -36,7 +36,6 @@ const (
 	kurtosisLogLevelArg    = "kurtosis-log-level"
 
 )
-var defaultKurtosisLogLevel = logrus.InfoLevel.String()
 
 var SandboxCmd = &cobra.Command{
 	Use:   command_str_consts.SandboxCmdStr,
@@ -54,7 +53,7 @@ func init() {
 		&kurtosisLogLevelStr,
 		kurtosisLogLevelArg,
 		"l",
-		defaultKurtosisLogLevel,
+		defaults.DefaultCoreLogLevel.String(),
 		fmt.Sprintf(
 			"The log level that Kurtosis itself should log at (%v)",
 			strings.Join(logrus_log_levels.GetAcceptableLogLevelStrs(), "|"),